(Subjectively) improve copy in binary
Also replaced one thrown error with a console.log. The rest of the errors were not thrown, and since this file is run from the shell, it will be an edge case that it will ever be caught. It looks messy when it happens. So it now returns an exit code of 1 to indicate things went sideways.