Fix jshint errors according to new settings
Use stricter jshint settings

I added
 - unused: "vars" // warns when you define and never use your variables
 - nonbsp: true // warns about "non-breaking whitespace" characters
 - maxdepth: 8 // lets you control how nested do you want your blocks to be
 - quotmark: false // true = all quotation marks have to either be single quotes or double quotes per file
I tried to use this option with true, but I'm not sure how you like it that way. As here are quite a few use cases where single quotes would have to be escaped

I removed:
 - trailing: true // this option no longer exists in jshint
 - expr: true // this option no longer exists in jshint
 - sub: true // suppresses warnings about using [] notation when it can be expressed in dot notation: person['name'] vs. person.name

I also used maxlen: 200 for a while, but I'm not sure how the line breaks work for the docs, since I did not try that. I still think it would be a good option to use

Be aware, that a couple of options are deprecated and will be removed at some point. Instead jscs could be used to enforce styling warnings.