Log which query tried to run on a finished transaction
The "commit has been called on this transaction" error message usually indicates a problem with chaining of query promises within a transaction block. In more complex transaction blocks it may not be immediately clear which query is improperly chained and the current error message does not help much. This change attaches the rejected query (SQL) as the 'sql' property of the error. This makes it easier to identify a problem in an application.